Refactor IISAccessLogParser and tests
IISAccessLogParser contained a long regular expression which was hard to
understand and maintain. This commit splits the regex into a more understandable
format.

There is an extra `?` at the end of the regex because some of the examples
in the spec have extra fields and some do not. These fields are not relevant to
the end result but they need to be captured in the final `?` to keep the regex
functioning correctly.

I have also refactored the spec to be more readable. The first test
`it "can be constructed from string"` did not actually do anything!