Use the fork of ActiverecordSessionStore
This addresses CVE-2019-16782

There has been a [vulnerability in the wild][1] around session hijacks in Rack and
related frameworks for a while now, but this has been fixed in Rack and Rails for a
while now. There's a [fix for the upstream version of ActiverecordSessionStore
since late 2019][2], but this hasn't been merged yet.

We weren't aware of this issue until recently, as it's only [just been added to the
Ruby Advisory DB][3]

This uses a fork of the upstream gem, [as suggested in the original PR][4] to fix
the immediate issue.

[1]:https://nvd.nist.gov/vuln/detail/CVE-2019-16782
[2]:https://github.com/rails/activerecord-session_store/pull/151
[3]:https://github.com/rubysec/ruby-advisory-db/pull/462
[4]:https://github.com/rails/activerecord-session_store/pull/151#issuecomment-631705247